requirements.txt
pytest-django
pytest-pythonpath==0.7.3
requests
pillow
sorl-thumbnail
django==2.2.6
djangorestframework
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0